requirements_base.txt
# python=3.10

# formatting
black
ruff

# logging
wandb

# JAX
    # Types
    beartype
    jaxtyping
    # JAX Neural networks
    flax
    jraph
    distrax
    optax
    orbax